Add layout computed lengths and remove margins.
A box may be expressed in many units. They all eventually resolve to a pixel
value. This implements measurements of each unit according to the definition of
how each unit is measured.

Margins can be implemented by adding empty boxes between boxes. Margins are only
useful if they collapse (otherwise, they are just a padding that can be
implemented by having a subbox). However, collapsing margins are problematic and
hard to calculate if they are not of the same units. HTML export can figure out
what empty boxes mean.