Replace stacks with partitions.
Partitions is a superset of stacks. In addition to detecting horizontally or
vertically aligned boxes, partitions also find stacked boxes that are not
aligned, as well as groups of boxes within a stack.

Partitioning works by maximizing whitespace among non-overlapping boxes. The
assumption is that non-overlapping boxes have a visual hierarchy, and whitespace
runs along the entire width or height of the root to divide up the children.
When applied recursively, each level of the hierarchy may be found.